+++ src/main/java/com/example/demo/domains/product/entity/ProductDetailImg.java
@@ -12,7 +12,6 @@
 package com.example.demo.domains.product.entity;
 
 import jakarta.persistence.*;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -24,7 +23,7 @@ public class ProductDetailImg {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "productDetailImg_id")
-    private int id;
+    private long id;
 
     private String imageUrl;
